Используем стандартные типы в своем коде.

+++ src/mov_mp4.cpp
@@ -1,36 +1,36 @@
 /*
  * Copyright (C) 2015-2020 Anton Shekhovtsov
- * Copyright (C) 2023-2024 v0lt
+ * Copyright (C) 2023-2025 v0lt
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
 #include "mov_mp4.h"
 
-unsigned long MovParser::read4()
+uint32_t MovParser::read4()
 {
-	unsigned long r;
+	uint32_t r;
 	if (hfile) {
-		unsigned long w;
+		DWORD w;
 		ReadFile(hfile, &r, 4, &w, 0);
 	}
 	else {
-		r = *(unsigned long*)p; p += 4;
+		r = *(uint32_t*)p; p += 4;
 	}
 	offset += 4;
 	r = _byteswap_ulong(r);
 	return r;
 }
 
-__int64 MovParser::read8()
+int64_t MovParser::read8()
 {
-	__int64 r;
+	int64_t r;
 	if (hfile) {
-		unsigned long w;
+		DWORD w;
 		ReadFile(hfile, &r, 8, &w, 0);
 	}
 	else {
-		r = *(__int64*)p; p += 8;
+		r = *(int64_t*)p; p += 8;
 	}
 	offset += 8;
 	r = _byteswap_uint64(r);
